Add support for non-root user mode to the init image (#955)
* init: add non-root user mode support

The aws-for-fluent-bit init image creates configs and other resources in the /init directory (root filesystem).
If a customer wants to run a container (with this image), with non-root user configured, the init setup fails.

This commit adds a fallback logic. When write to /init fails, it falls back to /tmp/init.
It enables non-root user mode support.

* go.mod: update the module name

Currently, the go.mod only pulls in dependencies for the load_tests/validation module.
This update allows go.mod to pull in dependencies for other submodules like init.